requirements.txt
elevenlabs>=1.0.0
pyaudio>=0.2.13
python-dotenv>=1.0.0
